requirements.txt
Django==1.7.1
Pillow==2.6.1
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.1.1
image==1.3.3
numpy==1.7.0
psycopg2==2.5.4
pymongo==2.7.2
scipy==0.11.0
static3==0.5.1
wsgiref==0.1.2
scikit-learn==0.13.1